xene-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
cryptography==43.0.1
python-jose[cryptography]==3.3.0
supabase==2.7.4
apscheduler==3.10.4
feedparser==6.0.11
python-dotenv==1.0.1
pydantic==2.8.2
